// src/game.services/profile.services/profile.interface.ts
export interface ProfilePicture {
    userName: string;
    fileFormat: string;
    profilePicture: string;
    uploadedAt: number;
    fileSize: number;
    likes: PictureLikes[];
    id?: string
}


export interface PictureLikes {
    userName?: string;
    timestamp?: number;
    likeId: string; 
}